# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Recall API
  version: "1"
  description: |
    HTTP REST + SSE surface for the Recall Overwatch metrics app.

    Available in two flavors:
      * **Server binary** (`Recall-server`) — exposes this API by default
      * **Wails desktop binary with `--server`** — same API, same routes

    Conventions:
      * Every JSON endpoint is mounted under the `/api/v1/` prefix —
        the version is bumped on breaking changes.
      * Resources are nouns; sub-resources hang off the parent
        (`/matches/{match_key}/visibility`, `/settings/tesseract`).
      * Methods reflect intent: `GET` to read, `PUT` to replace/upsert,
        `DELETE` to clear or wipe, `POST` to start a server-side action
        that doesn't map to a single resource (the parse run).
      * Writes return `204 No Content` when there is no useful body
        to echo back; `202 Accepted` for actions whose effect is
        asynchronous (parse).
      * Errors come back as `application/problem+json` (RFC 9457) with
        the appropriate 4xx/5xx status. `type` is a stable URI under
        `https://github.com/sound-barrier/recall/problems/{slug}` that
        clients can switch on; body-validation problems carry a
        per-field `errors` array, and the data-update 422 carries a
        `failed_assets` list (RFC 9457 §3.2 extension members). The one
        exception is `405 Method Not Allowed`, emitted plain by the
        router.
...

...
    `RECALL_SERVER_ADDR`.
  license:
...
servers:
...
paths:
...
  /api/v1/exports:
    get:
      operationId: ExportData
      tags: [Backup]
      summary: Download a full backup of the local parsed-match database
      description: |
        Returns a snapshot of the local database in one of two formats:

          * `format=json` (default) — a single JSON document containing
            every per-screenshot row across all parent tables plus the
            screenshots-dirs lookup, wrapped in a schema-versioned
            envelope.
          * `format=csv` — a ZIP archive containing one CSV file per
            parent and child table plus a `manifest.json` with the
            envelope metadata. Same data version as the JSON variant
            (`schema = recall-export/v1`), different container chosen
            by users who want to inspect or edit the data in Excel /
            Sheets.

        Both round-trip through `POST /api/v1/imports` — the server
        sniffs the first bytes to detect the format.
      parameters:
        - in: query
          name: format
          required: false
          schema:
            type: string
            enum: [json, csv]
            default: json
          description: Wire format. Defaults to JSON.
      responses:
        "200":
          description: Export download.
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="recall-export-20260526-013000.json"'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecallExport"
            application/zip:
              schema:
                type: string
                format: binary
                description: A ZIP archive — returned when `format=csv`.
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/exports/bundle:
    post:
      operationId: ExportBundle
      tags: [Backup]
      summary: Build a selection-aware export bundle (ZIP)
      description: |
        Builds a `.zip` containing the existing `recall-export/v1`
        JSON (filtered to the requested matches) at `data.json`, every
        referenced screenshot under `screenshots/<filename>`, and a
        root `manifest.json` listing screenshot → match_key mappings
        for sanity-checking after restore.

        The `match_keys` body field carries the user's explicit
        selection. `include_unknown` adds every record whose
        `data.map` is empty (the same "unknown" definition the
        Matches view uses); `include_hidden` adds every record
        currently in `hidden_matches`. Toggles UNION onto the
        explicit selection — a request with `match_keys: []` and
        `include_hidden: true` exports every hidden match.

        Restore via `POST /api/v1/imports` — the `data.json` shape
        is the same `recall-export/v1` envelope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [match_keys]
              properties:
                match_keys:
                  type: array
                  items:
                    type: string
                  description: |
                    Explicit list of match_keys to include in the
                    bundle. Empty array combined with
                    `include_unknown=false` and
                    `include_hidden=false` produces an empty
                    bundle (manifest with `match_count: 0`).
                include_unknown:
                  type: boolean
                  default: false
                  description: |
                    Add every record whose `data.map` is empty (the
                    Matches view's "unknown" filter).
                include_hidden:
                  type: boolean
                  default: false
                  description: |
                    Add every record currently in `hidden_matches`.
      responses:
        "200":
          description: |
            ZIP bundle with the same `Content-Disposition: attachment;
            filename="recall-bundle-<timestamp>.zip"` header pattern
            as `GET /api/v1/exports`.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/imports:
    post:
      operationId: ImportData
      tags: [Backup]
      summary: Restore the local database from a previously-exported backup
      description: |
        Replaces every row in the local database with the contents of
        the uploaded payload. Validates the envelope's `schema` field
        before touching the store; payload-level parse failures
        (not JSON / not ZIP, malformed JSON, ZIP open failure) are
        rejected with `400`; semantic-validation failures
        (unsupported schema, missing required field, null array
        entries) are rejected with `409`. The existing data is left
        untouched in both rejection paths.

        Accepts BOTH container formats — the server sniffs the payload's
        first bytes to detect a JSON envelope (`{` ...) or a ZIP archive
        (`PK\x03\x04` magic) and routes accordingly.

        The operation is REPLACE, not merge: anything currently in the
        store is dropped after the payload is validated and before the
        import lands. Callers should surface a confirmation step.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecallExport"
          application/zip:
            schema:
              type: string
              format: binary
              description: A ZIP archive produced by `GET /api/v1/exports?format=csv`.
      responses:
        "204":
          description: Import applied.
        "400":
          description: |
            Payload was not parseable as JSON or ZIP (decode failure,
            ZIP-open failure, payload neither container format).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: |
            Payload was syntactically valid JSON / ZIP but failed
            semantic validation (missing required field on a row,
            unsupported schema version, null entry in an array, etc.).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status]
      description: |
        RFC 9457 (Problem Details for HTTP APIs) error body, returned as
        `application/problem+json` for every 4xx/5xx the API surfaces. `type`
        is a stable URI under
        `https://github.com/sound-barrier/recall/problems/` that clients can
        switch on; `instance` is the request path. Members beyond those listed
        here are RFC 9457 §3.2 extension members (the schema stays open).
      properties:
        type:
          type: string
          format: uri
          description: A URI identifying the problem class.
          example: https://github.com/sound-barrier/recall/problems/invalid-body
        title:
          type: string
          description: Short, human-readable summary of the problem class.
          example: Bad Request
        status:
          type: integer
          description: The HTTP status code, duplicated in the body for convenience.
          example: 400
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: The request path that produced the problem.
        errors:
          type: array
          description: |
            §3.2 extension — per-field validation failures, present on a
            malformed request body.
          items:
            $ref: "#/components/schemas/FieldError"
        failed_assets:
          type: array
          description: |
            §3.2 extension — the asset(s) whose SHA-256 sidecar verification
            failed (the data-update 422).
          items:
            type: string
    FieldError:
      type: object
      required: [field, detail]
      description: One offending request field and why it was rejected.
      properties:
        field:
          type: string
          description: The request-body field name.
        detail:
          type: string
          description: Why the field was rejected.
...
    RecallExport:
      type: object
      # `schema` is the only field the import path strictly requires
      # — it gates the version check. `exported_at` and
      # `recall_version` are metadata-only echoes; both can be absent
      # on hand-edited or older-format payloads without breaking the
      # import.
      required: [schema]
      description: |
        Schema-versioned envelope for the local backup format produced
        by `GET /api/v1/exports` and consumed by `POST /api/v1/imports`.
        The per-row arrays mirror the SQLite parent + child tables;
        the `screenshots_dirs` map carries the path-by-id reference so
        FKs survive the auto-increment reset that happens during
        import.
      properties:
        schema:
          type: string
          description: Envelope version. Current value `recall-export/v1`.
          example: recall-export/v1
        exported_at:
          # Stored verbatim — the import never parses this string, it
          # only echoes back in any "import succeeded at" UI surface.
          # No format constraint so an old export from a different
          # build (or a hand-edited payload) imports cleanly even if
          # its timestamp shape isn't strictly RFC 3339. `null` is
          # also accepted (= unknown export time).
          type: string
        recall_version:
          # Optional metadata — `null` = unknown source version,
          # equivalent to omitting the field.
          type: string
          example: 0.1.4
        screenshots_dirs:
          # `null` is treated as an empty map — semantically equivalent
          # to `{}`. ImportData iterates a nil map without effect.
          type: object
          additionalProperties:
            type: string
          description: source-side `id` (stringified) → screenshots folder path
          example:
            "1": /Users/jacob/Documents/Overwatch/Screenshots
        # `null` is treated as an empty list — equivalent to `[]`. A
        # missing field decodes the same way. The server still rejects
        # individual `null` entries inside the array (each row must
        # carry a non-empty filename); see ImportData's loop in
        # pkg/app/export.go.
        summaries:
          type: array
          items:
            $ref: "#/components/schemas/SummaryExportRow"
        teams:
          type: array
          items:
            $ref: "#/components/schemas/TeamsExportRow"
        personals:
          type: array
          items:
            $ref: "#/components/schemas/PersonalExportRow"
        ranks:
          type: array
          items:
            $ref: "#/components/schemas/RankExportRow"
        unknowns:
          type: array
          items:
            $ref: "#/components/schemas/UnknownExportRow"
    SummaryExportRow:
      type: object
      additionalProperties: true
      required: [filename, match_key]
      description: |
        One row of `summary_screenshots` plus its `summary_heroes_played`
        child rows. Mirrors `pkg/db/store.go::SummaryRow`. Import-side
        is permissive (extra keys ignored); export-side emits every
        listed field.
      properties:
        filename:        { type: string }
        match_key:       { type: string }
        parsed_at:       { type: string }
        screenshots_dir_id: { type: integer, minimum: 1, description: "FK into `screenshots_dirs.id`. Never zero — the sentinel row at id=1 stands in for 'dir unset at parse time' (e.g. a row imported with an orphan FK or a test fixture that didn't supply a path)." }
        map:             { type: string }
        map_raw:         { type: string, description: "Raw OCR text when the canonical map lookup failed." }
        playlist:            { type: string }
        hero:            { type: string }
        hero_raw:        { type: string }
        result:          { type: string }
        final_score:     { type: string }
        date:            { type: string }
        finished_at:     { type: string }
        game_length:     { type: string }
        perf_elim_total:               { type: integer }
        perf_elim_avg_per_10min:       { type: number }
        perf_assists_total:            { type: integer }
        perf_assists_avg_per_10min:    { type: number }
        perf_deaths_total:             { type: integer }
        perf_deaths_avg_per_10min:     { type: number }
        heroes_played:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              hero:           { type: string }
              percent_played: { type: integer }
              play_time:      { type: string }
    TeamsExportRow:
      type: object
      additionalProperties: true
      required: [filename, match_key]
      description: |
        One row of `teams_screenshots` plus its
        `teams_hero_stats` child rows. Mirrors
        `pkg/db/store.go::TeamsRow`.
      properties:
        filename:        { type: string }
        match_key:       { type: string }
        parsed_at:       { type: string }
        screenshots_dir_id: { type: integer, minimum: 1, description: "FK into `screenshots_dirs.id`. Never zero — see SummaryExportRow." }
        map:             { type: string }
        map_raw:         { type: string }
        playlist:            { type: string }
        hero:            { type: string }
        hero_raw:        { type: string }
        eliminations:    { type: integer }
        assists:         { type: integer }
        deaths:          { type: integer }
        damage:          { type: integer }
        healing:         { type: integer }
        mitigation:      { type: integer }
        hero_stats:
          type: array
          items:
            $ref: "#/components/schemas/HeroStatExportRow"
    PersonalExportRow:
      type: object
      additionalProperties: true
      required: [filename, match_key]
      description: |
        One row of `personal_screenshots` plus its `personal_hero_stats`
        child rows. Mirrors `pkg/db/store.go::PersonalRow`.
      properties:
        filename:        { type: string }
        match_key:       { type: string }
        parsed_at:       { type: string }
        screenshots_dir_id: { type: integer, minimum: 1, description: "FK into `screenshots_dirs.id`. Never zero — see SummaryExportRow." }
        hero:            { type: string }
        hero_raw:        { type: string }
        hero_stats:
          type: array
          items:
            $ref: "#/components/schemas/HeroStatExportRow"
    RankExportRow:
      type: object
      additionalProperties: true
      required: [filename, match_key]
      description: |
        One row of `rank_screenshots` plus its `rank_modifiers` +
        `rank_sr` child rows. Mirrors `pkg/db/store.go::RankRow`.
      properties:
        filename:        { type: string }
        match_key:       { type: string }
        parsed_at:       { type: string }
        screenshots_dir_id: { type: integer, minimum: 1, description: "FK into `screenshots_dirs.id`. Never zero — see SummaryExportRow." }
        rank:            { type: string }
        level:           { type: integer }
        rank_progress:   { type: integer }
        change_percent:  { type: integer }
        result:          { type: string }
        modifiers:
          type: array
          items: { type: string }
        sr:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              hero:   { type: string }
              sr:     { type: integer }
              change: { type: integer }
    UnknownExportRow:
      type: object
      additionalProperties: true
      required: [filename, match_key]
      description: |
        One row of `unknown_screenshots`. Mirrors
        `pkg/db/store.go::UnknownRow`. No domain columns — kept so
        parse runs never drop a screenshot on the floor.
      properties:
        filename:        { type: string }
        match_key:       { type: string }
        parsed_at:       { type: string }
        screenshots_dir_id: { type: integer, minimum: 1, description: "FK into `screenshots_dirs.id`. Never zero — see SummaryExportRow." }
    HeroStatExportRow:
      type: object
      additionalProperties: true
      required: [hero, stat_key, stat_value]
      description: |
        One `(hero, stat_key, stat_value)` row from
        `teams_hero_stats` or `personal_hero_stats`.
      properties:
        hero:       { type: string }
        stat_key:   { type: string }
        stat_value: { type: integer }
...
  responses:
    BadRequest:
      description: Malformed request body or query parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    InternalError:
      description: Unhandled server-side error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...